Fix compiler warning

+++ ares/ares__read_line.c
@@ -46,7 +46,7 @@ int ares__read_line(FILE *fp, char **buf, size_t *bufsize)
 
   while (1)
     {
-      if (!fgets(*buf + offset, *bufsize - offset, fp))
+      if (!fgets(*buf + offset, (int)(*bufsize - offset), fp))
         return (offset != 0) ? 0 : (ferror(fp)) ? ARES_EFILE : ARES_EOF;
       len = offset + strlen(*buf + offset);
       if ((*buf)[len - 1] == '\n')
